fix: 좌표 관련 전부 double type으로 변경

+++ back-end/plogging/src/main/java/com/backend/plogging/dto/request/plogging/PloggingPostRequestDto.java
@@ -13,10 +13,10 @@ import javax.persistence.Column;
 public class PloggingPostRequestDto {
 
     private Float distance;
-    private Float startLat;
-    private Float startLng;
-    private Float endLat;
-    private Float endLng;
+    private Double startLat;
+    private Double startLng;
+    private Double endLat;
+    private Double endLng;
     private Float runningTime;
 
 }
